requirements.txt
grpcio==1.16.0
pkg-resources==0.0.0
protobuf==3.17.3
six==1.16.0